qcacld-3.0: Remove obsolete csr_msg_processor() code
Currently csr_msg_processor() contains logic that is conditionally
compiled based upon ANI_RTT_DEBUG and ANI_EMUL_ASSOC. These are
obviously obsolete references since these are the only references in
the code, and since they are referencing a variable that is not
present in the function. Since these are obsolete, remove them.

Change-Id: Ie258f188dd2791be96f2e95fcf4e8a71c4ea44eb
CRs-Fixed: 2103630
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>