qcacmn: Flush memory table on debugfs iterator stop
When printing memory allocations via debugfs, the dedupe table is
flushed to output when it reaches capacity. If the dedupe table is not
full when the allocation enumeration is complete, the entries contained
will be dropped. Flush the dedupe table to output one last time when
enumeration is complete to avoid dropping memory allocations from the
output.

Change-Id: I10f866b7bfa13c871d75e3e4ff798a8a820482df
CRs-Fixed: 2110106
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>