qcacld-3.0: Fix potential infinite loop in get_container_ies_len
In the function get_container_ies_len, len (uint32) is calculated
from the length of the buffer parsed. Then it is copied to the uint8
pnConsumed pointer from the calling function. This could lead to
pnConsumed becoming 0 if len exceeds 255 and would cause infinte
loop in the function unpack_core.

Add changes to make pnConsumed passed from unpack_core to be uint32
so that there are no issues in get_container_ies_len

Change-Id: Ia5770b4becf7dd1cf7cb97ec2e0d94f3c5f4ed54
CRs-Fixed: 2101200
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>