qcacld-3.0: Remove pStaWapiReqIE from the SAP context
Currently the SAP has logic to extract and locally save the WAPI
IE. But this is legacy code inherited from a version of the driver
that supported in-driver WAPI, and now the SAP WAPI functionality is
handled by hostapd. Since this local copy of the WAPI IE is never
used, remove it.

Change-Id: I57cad0d8e44ff1ca5a7a61454a21aac61bf873c0
CRs-Fixed: 2114153
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>