qcacld-3.0: Change return value of cds_init()
Currently cds_init() returns a pointer to the CDS Global Context.  At
one time this was necessary since it was required to subsequently pass
this context pointer to other CDS APIs. But now CDS internally
maintains a pointer to its context, and no CDS APIs now require this
pointer, so no longer return the pointer. The ultimate goal is to keep
all references to the CDS context inside CDS.

Change-Id: Id068a2d351e492eca7b84ef23f277939104c8b46
CRs-Fixed: 2117004
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>