qcacld-3.0: Differentiate roaming disabled by supplicant and on iface
Currently, do_not_roam flag is used for denoting both supplicant
disabled roaming and the roaming disabled on an interface during
the interface bring up. Differentiate these two flags with proper
nomenclature to make it clear and distinguishable. Name the new
flags to be supplicant_disabled_roaming and roaming_allowed_on_iface.

Currently in a sta+sta scenario roaming is supported only on wlan0.
Based on the flag roaming_allowed_on_iface, send/discard the
roaming commands to be sent to the firmware.

Change-Id: I1bd19e938b00c26c06ff06c73754560d4502ef2c
CRs-Fixed: 2130204
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>