qcacld-3.0: Do not destroy rx_buff_list_lock if it is not used
When CONFIG_HL_SUPPORT is defined, htt_rx_dbg_rxbuf_init is not called,
and rx_buff_list_lock is not initialized. rx_buff_list_lock should not
be destroyed if it is not initialized.

Do not call htt_rx_dbg_rxbuf_deinit when CONFIG_HL_SUPPORT is defined.

Change-Id: I9d64313efd5a9978b3fb953ac5d7ffb466abe010
CRs-Fixed: 2134345
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>