qcacld-3.0: Fix memory leak while handling deauth req's error case
When caller calls lim_process_mlm_deauth_req() API, it checks for
session's validity before passing it forward. If session is found
to be invalid then it is returning immediately without releasing the
memory. Caller is expecting this API to release the memory in any
error cases.

Release the memory before returning to caller in error case.

CRs-Fixed: 2138983
Change-Id: I800d717cd22a6e2731ef211743e4844852a132a3
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>