qcacld-3.0: Remove unused hdd_roam_register_tdlssta() param
The parameters to function hdd_roam_register_tdlssta() currently
includes ucastSig. This "signature" parameter dates back to a previous
version of the driver that used signatures to synchronize the data
plane with the control plane. However the current version of the
driver does not use this parameter, so remove it as part of the plan
to completely remove these signatures from throughout the driver.

Change-Id: Iee51d79e58501e3e8312c6c7a21b98c1ae6bf268
CRs-Fixed: 2200886
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>