qcacld-3.0: Return error if channel list is NULL
Hostapd sends channel list as NULL in do_acs request
when country mode is set to world and hw mode is 11a.
This results in scan to be triggered on all 5g channels
and driver to choose 5g channel as best channel.

Fix is to return error to hostapd if channel list
received is NULL.

Change-Id: I17aacfdbeea9803023a66f54de7b596612ea97c3
CRs-Fixed: 2192858
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>