qcacld-3.0: Relocate hdd_statistics_cb()
Function hdd_statistics_cb() is registered as a callback function by
__iw_get_statistics() when calling sme_get_statistics(). Since
hdd_statistics_cb() and __iw_get_statistics() are tightly coupled it
makes sense for them to be co-located, but currently they are
not. Therefore relocate hdd_statistics_cb() to be adjacent to
__iw_get_statistics(). Note that this is a precursor to further
cleanup of this pair of functions.

Change-Id: Ie36d107d9ef2e541444c5fb9d7bfba7e9e82c1e6
CRs-Fixed: 2201907
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>