qcacld-3.0: Set wmi_handle in pdev before registering EGAP event handler
wmi_handle is needed while registering for wmi event handler. But it is
not set for pdev before registering for EGAP event handler. With this,
EGAP support remains disabled which may end up in increased power
consumption.

Move setting of wmi_handle in pdev before registering for EGAP event.

Change-Id: Icb8accc738a3aaad7fbc1201c7c0d0b77945e8f2
CRs-Fixed: 2203380
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>