qcacld-3.0: Dont process driver unload during system reboot
During the system reboot when the firmware goes offline before the
driver unload there is no way to detect whether the it is because
of the reboot or a assert in the firmware. So, the driver waits
for the firmware to recover before continuing with unloading
the device.

Since this is a reboot case the firmware will not respawn the
execution and host driver is struck.

To mitigate the issue register a reboot notifier with the kernel
and if reboot is triggered proceed with a fake unload.

Change-Id: I33d14c393930e9ee737d38ffdb26e921fb057f29
CRs-Fixed: 2211776
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>