qcacld-3.0: Pass correct param type via QoS callback
The QoS callback is supposed to pass a mac_handle_t (aka tHalHandle)
to HDD, but in most cases what is actually passed is a tpAniSirGlobal.
Even though these ultimately reference the same structure, an upcoming
change will enforce type safety so update the bad callers to pass a
parameter of the correct type.

Change-Id: I5913c3cf577606f8c207b1e40897b8cbe681de48
CRs-Fixed: 2267442
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>