qcacld-3.0: Update channel in DS params and HT Info IE after CSA
After CSA, channel is not getting updated in DS params and HT info
IE in BSS descriptor of corresponding session. As a result, channel
in cfg80211 is still old due to which freq getting displayed in UI
is still the older one.

Update channel in DS params and HT info IE in BSS descriptor after CSA.

Change-Id: I4a0f301ccd6155dc459fa1bfa4fbd0c59c04e0f8
CRs-Fixed: 2244619
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>