qcacld-3.0: Move objmgr create/delete from HDD->WMA
Currently the driver creates and deletes the objmgr peer
in HDD layer. According to the latest requirement, the CDP
peer create/delete needs the pointer of objmgr peer, which
thus has to created before CDP peer create, and deleted
after CDP peer delete, which is done in WMA layer.

Remove the objmgr peer create/delete from HDD, and move it
to the WMA layer.

Change-Id: I5bd62a83864e09a3728109e2deec2adf4bb899b5
CRs-Fixed: 2264892
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>