qcacmn: Use statically allocated memory for a few DFS structures
Change a few memory allocations of the DFS module from dynamic to
statically allocated memory from .bss segment to avoid any possible
runtime memory allocation failure because of the requested contiguous
memory of a large size.

The static allocation will not work for Multiple RADIO based boards as
it will use the same memory for all the radios.

Change-Id: I93b30a0a3f2f7d91466c930c12e9fae60699605a
CRs-Fixed: 2302097
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>